Runner on .net standard 2.1
