--- src/DataMasker/Interfaces/IDataGenerator.cs
@@ -1,13 +1,0 @@
-﻿using Bogus.DataSets;
-using DataMasker.Models;
-
-namespace DataMasker.Interfaces
-{
-    public interface IDataGenerator
-    {
-        object GetValue(
-            ColumnConfig columnConfig,
-            object existingValue,
-            Name.Gender? gender);
-    }
-}
