+++ src/DataMasker/Models/SqlValueConfig.cs
@@ -1,0 +1,11 @@
+
+namespace DataMasker.Models
+{
+  public class SqlValueConfig
+  {
+    /// <summary>
+    /// SQL Query, the first column of the first row will be selected
+    /// </summary>
+    public string Query { get; set; }
+  }
+}
